moved make_privkey_address to tests.utils
